% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcode.R
\name{parse_postcode}
\alias{parse_postcode}
\title{Extract postcode from an address}
\usage{
parse_postcode(address)
}
\arguments{
\item{address}{Address variable}

\item{.data}{Input data set}
}
\description{
Extracts postcode (5 digits). If more than one postcode (5 digits) can be found, it will return the last postcode found.
}
\examples{
parse_postcode("5, shah alam, 40150")
}
