% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcode.R
\name{parse_postcode}
\alias{parse_postcode}
\title{Extract postcode(s) from an address}
\usage{
parse_postcode(.data, address)
}
\arguments{
\item{.data}{Input data set}

\item{address}{Address variable}
}
\value{
A tibble with the number postcodes found (5 digits) & the extracted postcode
}
\description{
Extract postcode(s) from an address
}
